requirements.txt
black
pylint
radon
langchain-google-genai
google-cloud-bigquery
google-cloud-bigquery-storage
streamlit
langchain
langchain_groq
python-dotenv
tabulate>=0.9.0
SQLAlchemy
db-dtypes